🚚 api: move controller to src/team context
