+++ api/src/team/application/organization-invitations/organization-invitation.controller.js
@@ -106,6 +106,29 @@ const sendInvitations = async function (request, h) {
   return h.response(organizationInvitationSerializer.serialize(organizationInvitations)).created();
 };
 
+/**
+ *
+ * @param request
+ * @param h
+ * @param dependencies
+ * @returns {Promise<any>}
+ */
+const sendInvitationByLangAndRole = async function (request, h, dependencies = { organizationInvitationSerializer }) {
+  const organizationId = request.params.id;
+  const invitationInformation =
+    await dependencies.organizationInvitationSerializer.deserializeForCreateOrganizationInvitationAndSendEmail(
+      request.payload,
+    );
+
+  const organizationInvitation = await usecases.createOrganizationInvitationByAdmin({
+    organizationId,
+    email: invitationInformation.email,
+    locale: invitationInformation.lang,
+    role: invitationInformation.role,
+  });
+  return h.response(dependencies.organizationInvitationSerializer.serialize(organizationInvitation)).created();
+};
+
 export const organizationInvitationController = {
   acceptOrganizationInvitation,
   findPendingInvitations,
@@ -113,4 +136,5 @@ export const organizationInvitationController = {
   getOrganizationInvitation,
   sendInvitations,
   sendScoInvitation,
+  sendInvitationByLangAndRole,
 };
